+++ rebook-security/src/main/java/com/be/rebook/security/controller/JoinController.java
@@ -1,8 +1,8 @@
-package com.be.rebook.domain.security.controller;
+package com.be.rebook.security.controller;
 
 import com.be.rebook.domain.members.entity.Members;
-import com.be.rebook.domain.security.service.JoinService;
-import com.be.rebook.domain.security.dto.JoinDTO;
+import com.be.rebook.security.service.JoinService;
+import com.be.rebook.security.dto.JoinDTO;
 import com.be.rebook.global.config.BaseResponse;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.ResponseEntity;
