+++ rebook-security/src/main/java/com/be/rebook/security/dto/CustomUserDetails.java
@@ -1,4 +1,4 @@
-package com.be.rebook.domain.security.dto;
+package com.be.rebook.security.dto;
 
 
 import com.be.rebook.domain.members.entity.Members;
