+++ rebook-security/src/main/java/com/be/rebook/security/jwt/CustomLogoutFilter.java
@@ -1,6 +1,6 @@
-package com.be.rebook.domain.security.jwt;
+package com.be.rebook.security.jwt;
 
-import com.be.rebook.domain.security.repository.RefreshTokensRepository;
+import com.be.rebook.security.repository.RefreshTokensRepository;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.ServletRequest;
