+++ rebook-security/src/main/java/com/be/rebook/security/service/JoinService.java
@@ -1,9 +1,9 @@
-package com.be.rebook.domain.security.service;
+package com.be.rebook.security.service;
 
 import com.be.rebook.domain.members.entity.Members;
 import com.be.rebook.domain.members.repository.MembersRepository;
-import com.be.rebook.domain.security.dto.JoinDTO;
-import com.be.rebook.domain.security.utility.InputVerifier;
+import com.be.rebook.security.dto.JoinDTO;
+import com.be.rebook.security.utility.InputVerifier;
 import com.be.rebook.global.exception.BaseException;
 import com.be.rebook.global.exception.ErrorCode;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
