+++ rebook-security/src/main/java/com/be/rebook/security/utility/InputVerifier.java
@@ -1,4 +1,4 @@
-package com.be.rebook.domain.security.utility;
+package com.be.rebook.security.utility;
 
 public class InputVerifier {
     private static final String[] SQL_INJECTION_CHARS = {"'", "\"", ";", "--", "\\"};
